ml-service/requirements.txt
fastapi
uvicorn
pandas
numpy
scikit-learn
joblib
mysql-connector-python